+++ tests/jupyter/conftest.py
@@ -1,3 +1,5 @@
+import os
+import subprocess
 from dataclasses import dataclass
 from pathlib import Path
 
@@ -11,10 +13,17 @@ class DummyProject:
     expected_manager: str
     working_dir: Path
 
-    def verify_system_paths(self) -> None:
+    def verify_project(self) -> None:
         helper = VenvHelper()
         assert helper.exec_prefix.startswith(str(self.working_dir))
         assert helper.executable.startswith(str(self.working_dir))
+
+        self._assert_packages_are_installed(helper.executable)
+
+    @classmethod
+    def _assert_packages_are_installed(cls, python: str) -> None:
+        subprocess.check_call([python, "-c", "import pytest"])
+        subprocess.check_call([python, "-c", "import ipykernel"])
 
 
 @pytest.fixture
@@ -22,25 +31,30 @@ def dummy_project(request: pytest.FixtureRequest, tmp_path: Path, monkeypatch: p
     manager = request.param
 
     monkeypatch.chdir(tmp_path)
-    tmp_path.joinpath(f"{manager}.lock").write_text(_UV_LOCK)
     tmp_path.joinpath("pyproject.toml").write_text(_PYPROJECT)
+    # tmp_path.joinpath("src/").write_text()
+
+    env = {**os.environ}
+    env.pop("VIRTUAL_ENV", None)
+
+    subprocess.check_output([manager, "lock"], env=env)
+    assert tmp_path.joinpath(f"{manager}.lock").is_file()
+    subprocess.check_output([manager, "sync"], env=env)
 
     return DummyProject(manager, tmp_path)
 
 
-_PYPROJECT = """
+_PYPROJECT = f"""# See {__file__}
 [project]
 name = "dummy-project"
 version = "0.1.0"
-"""
 
-_UV_LOCK = """
-version = 1
-revision = 2
 requires-python = ">=3.11"
 
-[[package]]
-name = "dummy-project"
-version = "0.1.0"
-source = { virtual = "." }
+dependencies = [
+    "pytest == {pytest.__version__}",
+]
+
+[tool.poetry]
+package-mode = false
 """